hashmap

+++ Taller 1 - Hamburguesas/src/logica/Combo.java
@@ -24,7 +24,7 @@ public class Combo implements Producto {
 
 	}
 	
-	public double getPrecio() {
+	public int getPrecio() {
 		return descuento;
 
 	}
